new personal risk function
